requirements.txt
boto3
cryptography
pyjwt
python-dotenv